+++ silkopter/brain/src/sink/PIGPIO.h
@@ -58,7 +58,7 @@ private:
 
     std::vector<PWM_Channel> m_pwm_channels;
 
-    void set_pwm_value(size_t idx, float value);
+    void set_pwm_value(size_t idx, double value);
 };
 
 }
