+++ silkopter/libs/common/stream/IFloat.h
@@ -12,7 +12,7 @@ class IFloat : public IScalar_Stream<Type::FLOAT>
 public:
     typedef std::true_type can_be_filtered_t;
 
-    typedef float                       Value;
+    typedef double                       Value;
     typedef stream::Sample<Value>       Sample;
     virtual auto get_samples() const -> std::vector<Sample> const& = 0;
 };
