+++ silkopter/libs/common/stream/IVoltage.h
@@ -13,7 +13,7 @@ class IVoltage : public IScalar_Stream<Type::VOLTAGE>
 public:
     typedef std::true_type can_be_filtered_t;
 
-    typedef float                   Value; //volts
+    typedef double                   Value; //volts
     typedef stream::Sample<Value>     Sample;
     virtual auto get_samples() const -> std::vector<Sample> const& = 0;
 };
